unittests/ThunkLibs: Move utility code to a dedicated header
